Added random rocks
I added 2 or 3 rocks( randomly chosed the number and position) on the map.
The idea will be for the next pr to make them to take damage and when is attacked to spawn enemies. The rock already has a health property, it lacks the colision detaction, detaction of attack and  spawning the enemies. Hopefully i will know to implement them for the next pr :))