Added an Enemy!
The enemy has basic functionality, simply walking towards a player and performing an attack animation (player health is not added). The enemy can also be slain by a player attack (not bomb, the sword itself, which is dependent on attack direction). 
The functionality of the enemy object is handled in its own EnemyObject class, while the data of each object is kept in a list in the Engine class, sequentially calling their decision algorithms every frame.