Questalia - quest manager plugin
A plugin used to manage quests in TheAdventure game. Create active / inactive / completed quests with 2 different levels of aggressiveness: aggressive and neutral (they differ via the icon displayed)

Future directions: to display text, we would need a huge sprite with all the letters and just iterate over the chars of the text and display a new RenderableGameObject with the corresponding letter.